# storage/swagger.yaml
openapi: 3.0.0
info:
  title: 'OKR API'
  version: '1.0'
paths:
  /api/like:
    post:
      tags:
        - いいね追加
      description: コメントにいいねを追加する
      operationId: 8e1e9b2f8d273f462544126cfacae4a2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                comment_id:
                  description: コメントID
                  type: integer
                user_id:
                  description: ユーザーID
                  type: integer
              type: object
      responses:
        '200':
          description: いいね追加のDB更新
          content:
            application/json:
              schema:
                properties:
                  result: { description: 結果, type: boolean }
                type: object
  /api/like_remove:
    post:
      tags:
        - いいね取り消し
      description: コメントのいいねを取り消す
      operationId: 635a9eb0f75af7bfa21708a02e384b76
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                comment_id:
                  description: コメントID
                  type: integer
                user_id:
                  description: ユーザーID
                  type: integer
              type: object
      responses:
        '200':
          description: いいね取り消しDB更新
          content:
            application/json:
              schema:
                properties:
                  result: { description: 結果, type: boolean }
                type: object
  /api/okr/mine/get:
    post:
      tags:
        - 'OKR 取得処理'
      description: '自身の OKR を取得する'
      operationId: 9ae9bf8a073d08cb93d837b5ba125bbf
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                year:
                  description: 西暦
                  type: integer
                  example: '2022'
                quarter_id:
                  description: 四半期ID
                  type: integer
                  example: '1'
                is_archived:
                  description: アーカイブフラグ
                  type: integer
                  example: '0'
                is_include_full_year:
                  description: '通年 OKR 同時取得フラグ'
                  type: integer
                  example: '0'
              type: object
      responses:
        '200':
          description: 'OKR 取得成功'
          content:
            application/json:
              schema:
                properties:
                  objectives: { description: OKR, type: object }
                type: object
  /api/okr/ours/get:
    post:
      tags:
        - '会社全体 OKR 取得処理'
      description: 'ユーザに紐づく会社の全員の OKR を取得する'
      operationId: 850ca943ea684546a7af626563e36288
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                year:
                  description: 西暦
                  type: integer
                  example: '2022'
                quarter_id:
                  description: 四半期ID
                  type: integer
                  example: '1'
                user_id:
                  description: ユーザID
                  type: integer
                  example: '1'
                is_archived:
                  description: アーカイブフラグ
                  type: integer
                  example: '0'
                is_include_full_year:
                  description: '通年 OKR 同時取得フラグ'
                  type: integer
                  example: '0'
              type: object
      responses:
        '200':
          description: 'OKR 取得成功'
          content:
            application/json:
              schema:
                properties:
                  objectives: { description: OKR, type: object }
                type: object
  /api/quarter/get:
    post:
      tags:
        - 四半期取得処理
      description: 自身の会社に紐づく四半期情報を取得する
      operationId: 8de0fd211fe855c498362ae0794dc67e
      responses:
        '200':
          description: 四半期一覧取得成功
          content:
            application/json:
              schema:
                properties:
                  quarters: { description: 四半期, type: object }
                type: object